DRY-ed it a bit
A lot of parts of this program are redudant. I fixed some of the most obvious ones that had been bothering me for a while. Other components that can be reduced are ListaCitas and ListaExpedientes, and the register and edit components since they pretty much have the same forms, just different handling functions.